feat: Performance logging
