--- src/main/java/it/gov/pagopa/payhub/pdnd/anpr/c030/service/AnprC030ServiceConfig.java
@@ -1,11 +1,0 @@
-package it.gov.pagopa.payhub.pdnd.anpr.c030.service;
-
-import it.gov.pagopa.payhub.pdnd.config.pdnd.PdndServiceIntegratedConfig;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ConfigurationProperties(prefix = "app.pdnd.anpr.services.c030")
-public class AnprC030ServiceConfig extends PdndServiceIntegratedConfig {
-
-}
