+++ src/main/java/it/gov/pagopa/payhub/pdnd/anpr/service/AnprService.java
@@ -1,0 +1,8 @@
+package it.gov.pagopa.payhub.pdnd.anpr.service;
+
+import it.gov.pagopa.payhub.pdnd.dto.generated.Citizen;
+
+public interface AnprService {
+
+    Citizen getCitizenData(String fiscalCode);
+}
