--- src/main/java/it/gov/pagopa/payhub/pdnd/anpr/service/PdndService.java
@@ -1,8 +1,0 @@
-package it.gov.pagopa.payhub.pdnd.anpr.service;
-
-import it.gov.pagopa.payhub.pdnd.dto.generated.Citizen;
-
-public interface PdndService {
-
-    Citizen getCitizenData(String fiscalCode);
-}
