+++ src/main/java/it/gov/pagopa/payhub/pdnd/config/ApiClientConfig.java
@@ -1,0 +1,17 @@
+package it.gov.pagopa.payhub.pdnd.config;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@Data
+@SuperBuilder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ApiClientConfig {
+    private String baseUrl;
+    private int maxAttempts;
+    private long waitTimeMillis;
+    private boolean printBodyWhenError;
+}
