+++ src/main/java/it/gov/pagopa/payhub/pdnd/connector/pdnd/PdndService.java
@@ -1,0 +1,7 @@
+package it.gov.pagopa.payhub.pdnd.connector.pdnd;
+
+import it.gov.pagopa.payhub.pdnd.config.pdnd.PdndServiceIntegratedConfig;
+
+public interface PdndService {
+    String generateToken(PdndServiceIntegratedConfig pdndServiceIntegratedConfig);
+}
