+++ src/main/java/it/gov/pagopa/payhub/pdnd/connector/pdnd/client/PdndClient.java
@@ -1,7 +1,23 @@
 package it.gov.pagopa.payhub.pdnd.connector.pdnd.client;
 
+import it.gov.pagopa.payhub.pdnd.connector.pdnd.config.PdndApisHolder;
+import it.gov.pagopa.payhub.pdnd.connector.pdnd.generated.api.AuthApi;
 import it.gov.pagopa.payhub.pdnd.connector.pdnd.generated.dto.ClientCredentialsResponseDTO;
+import org.springframework.stereotype.Service;
 
-public interface PdndClient {
-  ClientCredentialsResponseDTO getAccessToken(String clientId, String clientAssertions);
+@Service
+public class PdndClient {
+
+  private static final String CLIENT_ASSERTION_TYPE = "urn:ietf:params:oauth:client-assertion-type:jwt-bearer";
+  private static final String GRANT_TYPE = "client_credentials";
+
+  private final AuthApi authApi;
+
+  public PdndClient(PdndApisHolder pdndApisHolder) {
+    authApi = pdndApisHolder.getBrokerEntityControllerApi();
+  }
+
+  public ClientCredentialsResponseDTO getAccessToken(String clientId, String clientAssertions) {
+    return authApi.createToken(clientAssertions, CLIENT_ASSERTION_TYPE, GRANT_TYPE, clientId);
+  }
 }
