--- src/test/java/it/gov/pagopa/payhub/pdnd/anpr/c003/service/AnprC003ServiceTest.java
@@ -1,61 +1,0 @@
-package it.gov.pagopa.payhub.pdnd.anpr.c003.service;
-
-import it.gov.pagopa.payhub.anpr.C003.dto.generated.*;
-import it.gov.pagopa.payhub.pdnd.anpr.c003.client.AnprC003ClientImpl;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.*;
-
-@ExtendWith(MockitoExtension.class)
-class AnprC003ServiceTest {
-
-    @Mock
-    private AnprC003ClientImpl anprC003ClientImpl;
-
-    @InjectMocks
-    private AnprC003Service anprC003Service;
-
-    @Test
-    void givenValidIdAnprAndFiscalCodeWhenGetUserDataThenReturnUserDataSuccessfully() {
-        String fiscalCode = "DNTCRL65S67M126K";
-        String idAnpr = "d20fcd8e-f228-323c-8924-6405b44879bf";
-
-        TipoInfoSoggettoEnte subTypeInfo = TipoInfoSoggettoEnte.builder()
-                .id("nome")
-                .chiave("Jed")
-                .valore(TipoInfoValore.S)
-                .valoreTesto("Nome del soggetto")
-                .valoreData("2024-11-02")
-                .dettaglio("")
-                .build();
-
-        TipoDatiSoggettiEnte subDataTypes = TipoDatiSoggettiEnte.builder()
-                .infoSoggettoEnte(List.of(subTypeInfo))
-                .build();
-
-        TipoListaSoggetti subTypeList = TipoListaSoggetti.builder()
-                .datiSoggetto(List.of(subDataTypes))
-                .build();
-
-        RispostaE002OK mockResponse = RispostaE002OK.builder()
-                .idOperazioneANPR("12345")
-                .listaSoggetti(subTypeList)
-                .listaAnomalie(null)
-                .build();
-
-        when(anprC003ClientImpl.getUserData(Mockito.any(RichiestaE002.class))).thenReturn(mockResponse);
-
-        RispostaE002OK response = anprC003Service.getUserData(idAnpr, fiscalCode);
-
-        assertNotNull(response);
-        assertEquals("Jed", response.getListaSoggetti().getDatiSoggetto().getFirst().getInfoSoggettoEnte().getFirst().getChiave());
-    }
-}
