--- src/test/java/it/gov/pagopa/payhub/pdnd/exception/PdndServicesExceptionHandlerTest.java
@@ -1,194 +1,0 @@
-package it.gov.pagopa.payhub.pdnd.exception;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import it.gov.pagopa.payhub.pdnd.config.json.JsonConfig;
-import jakarta.servlet.ServletException;
-import jakarta.validation.ConstraintViolationException;
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Pattern;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.http.MediaType;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.bean.override.mockito.MockitoSpyBean;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.ResultActions;
-import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.server.ServerErrorException;
-import org.springframework.web.servlet.mvc.method.annotation.RequestMappingHandlerAdapter;
-
-import java.time.LocalDateTime;
-import java.util.Set;
-
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.doThrow;
-
-@ExtendWith({SpringExtension.class})
-@WebMvcTest(value = {PdndServicesExceptionHandlerTest.TestController.class}, excludeAutoConfiguration = SecurityAutoConfiguration.class)
-@ContextConfiguration(classes = {
-        PdndServicesExceptionHandlerTest.TestController.class,
-        PdndServicesExceptionHandler.class,
-        JsonConfig.class})
-class PdndServicesExceptionHandlerTest {
-
-    public static final String DATA = "data";
-    public static final TestRequestBody BODY = new TestRequestBody("bodyData", null, "abc", LocalDateTime.now());
-
-    @Autowired
-    private MockMvc mockMvc;
-    @Autowired
-    private ObjectMapper objectMapper;
-
-    @MockitoSpyBean
-    private TestController testControllerSpy;
-    @MockitoSpyBean
-    private RequestMappingHandlerAdapter requestMappingHandlerAdapterSpy;
-
-    @RestController
-    @Slf4j
-    static class TestController {
-        @PostMapping(value = "/test", produces = MediaType.APPLICATION_JSON_VALUE)
-        String testEndpoint(@RequestParam(DATA) String data, @Valid @RequestBody TestRequestBody body) {
-            return "OK";
-        }
-    }
-
-    @Data
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class TestRequestBody {
-        @NotNull
-        private String requiredField;
-        private String notRequiredField;
-        @Pattern(regexp = "[a-z]+")
-        private String lowerCaseAlphabeticField;
-        private LocalDateTime dateTimeField;
-    }
-
-    private ResultActions performRequest(String data, MediaType accept) throws Exception {
-        return performRequest(data, accept, objectMapper.writeValueAsString(PdndServicesExceptionHandlerTest.BODY));
-    }
-
-    private ResultActions performRequest(String data, MediaType accept, String body) throws Exception {
-        MockHttpServletRequestBuilder requestBuilder = MockMvcRequestBuilders.post("/test")
-                .param(DATA, data)
-                .accept(accept);
-
-        if (body != null) {
-            requestBuilder
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content(body);
-        }
-
-        return mockMvc.perform(requestBuilder);
-    }
-
-    @Test
-    void handleMissingServletRequestParameterException() throws Exception {
-
-        performRequest(null, MediaType.APPLICATION_JSON)
-                .andExpect(MockMvcResultMatchers.status().isBadRequest())
-                .andExpect(MockMvcResultMatchers.jsonPath("$.code").value("PDND_SERVICES_BAD_REQUEST"))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.message").value("Required request parameter 'data' for method parameter type String is not present"));
-
-    }
-
-    @Test
-    void handleRuntimeExceptionError() throws Exception {
-        doThrow(new RuntimeException("Error")).when(testControllerSpy).testEndpoint(DATA, BODY);
-
-        performRequest(DATA, MediaType.APPLICATION_JSON)
-                .andExpect(MockMvcResultMatchers.status().isInternalServerError())
-                .andExpect(MockMvcResultMatchers.jsonPath("$.code").value("PDND_SERVICES_GENERIC_ERROR"))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.message").value("Error"));
-    }
-
-    @Test
-    void handleGenericServletException() throws Exception {
-        doThrow(new ServletException("Error"))
-                .when(requestMappingHandlerAdapterSpy).handle(any(), any(), any());
-
-        performRequest(DATA, MediaType.APPLICATION_JSON)
-                .andExpect(MockMvcResultMatchers.status().isInternalServerError())
-                .andExpect(MockMvcResultMatchers.jsonPath("$.code").value("PDND_SERVICES_GENERIC_ERROR"))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.message").value("Error"));
-    }
-
-    @Test
-    void handle4xxHttpServletException() throws Exception {
-        performRequest(DATA, MediaType.parseMediaType("application/hal+json"))
-                .andExpect(MockMvcResultMatchers.status().isNotAcceptable())
-                .andExpect(MockMvcResultMatchers.jsonPath("$.code").value("PDND_SERVICES_BAD_REQUEST"))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.message").value("No acceptable representation"));
-    }
-
-    @Test
-    void handleUrlNotFound() throws Exception {
-        mockMvc.perform(MockMvcRequestBuilders.post("/NOTEXISTENTURL"))
-                .andExpect(MockMvcResultMatchers.status().isNotFound())
-                .andExpect(MockMvcResultMatchers.jsonPath("$.code").value("PDND_SERVICES_NOT_FOUND"))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.message").value("No static resource NOTEXISTENTURL."));
-    }
-
-    @Test
-    void handleNoBodyException() throws Exception {
-        performRequest(DATA, MediaType.APPLICATION_JSON, null)
-                .andExpect(MockMvcResultMatchers.status().isBadRequest())
-                .andExpect(MockMvcResultMatchers.jsonPath("$.code").value("PDND_SERVICES_BAD_REQUEST"))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.message").value("Required request body is missing"));
-    }
-
-    @Test
-    void handleInvalidBodyException() throws Exception {
-        performRequest(DATA, MediaType.APPLICATION_JSON,
-                "{\"notRequiredField\":\"notRequired\",\"lowerCaseAlphabeticField\":\"ABC\"}")
-                .andExpect(MockMvcResultMatchers.status().isBadRequest())
-                .andExpect(MockMvcResultMatchers.jsonPath("$.code").value("PDND_SERVICES_BAD_REQUEST"))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.message").value("Invalid request content: lowerCaseAlphabeticField: must match \"[a-z]+\"; requiredField: must not be null"));
-    }
-
-    @Test
-    void handleNotParsableBodyException() throws Exception {
-        performRequest(DATA, MediaType.APPLICATION_JSON,
-                "{\"notRequiredField\":\"notRequired\",\"dateTimeField\":\"2025-02-05\"}")
-                .andExpect(MockMvcResultMatchers.status().isBadRequest())
-                .andExpect(MockMvcResultMatchers.jsonPath("$.code").value("PDND_SERVICES_BAD_REQUEST"))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.message").value("Cannot parse body: dateTimeField: Text '2025-02-05' could not be parsed at index 10"));
-    }
-
-    @Test
-    void handle5xxHttpServletException() throws Exception {
-        doThrow(new ServerErrorException("Error", new RuntimeException("Error")))
-                .when(requestMappingHandlerAdapterSpy).handle(any(), any(), any());
-
-        performRequest(DATA, MediaType.APPLICATION_JSON)
-                .andExpect(MockMvcResultMatchers.status().isInternalServerError())
-                .andExpect(MockMvcResultMatchers.jsonPath("$.code").value("PDND_SERVICES_GENERIC_ERROR"))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.message").value("500 INTERNAL_SERVER_ERROR \"Error\""));
-    }
-
-    @Test
-    void handleViolationException() throws Exception {
-        doThrow(new ConstraintViolationException("Error", Set.of())).when(testControllerSpy).testEndpoint(DATA, BODY);
-
-        performRequest(DATA, MediaType.APPLICATION_JSON)
-                .andExpect(MockMvcResultMatchers.status().isBadRequest())
-                .andExpect(MockMvcResultMatchers.jsonPath("$.code").value("PDND_SERVICES_BAD_REQUEST"))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.message").value("Error"));
-    }
-}
